refactor: add chatID parameter to commands for callbacks
A command can be both executed from a command or a callback coming from a button for example. Therefore, a check is made at the root execution to detect which entity was received in order to parse a chat ID that is defined in both situations. This prevents nil values from non-existing messages or callbacks.